requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.2
regex==2024.11.6
requests==2.32.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
